+++ src/anemoi/graphs/nodes/__init__.py
@@ -1,8 +1,20 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
 from .builders.from_file import LimitedAreaNPZFileNodes
 from .builders.from_file import NPZFileNodes
 from .builders.from_file import ZarrDatasetNodes
 from .builders.from_healpix import HEALPixNodes
 from .builders.from_healpix import LimitedAreaHEALPixNodes
+from .builders.from_icon import ICONCellGridNodes
+from .builders.from_icon import ICONMultimeshNodes
+from .builders.from_icon import ICONNodes
 from .builders.from_refined_icosahedron import HexNodes
 from .builders.from_refined_icosahedron import LimitedAreaHexNodes
 from .builders.from_refined_icosahedron import LimitedAreaTriNodes
@@ -20,4 +32,7 @@ __all__ = [
     "LimitedAreaTriNodes",
     "LimitedAreaHexNodes",
     "StretchedTriNodes",
+    "ICONMultimeshNodes",
+    "ICONCellGridNodes",
+    "ICONNodes",
 ]